+++ library/src/main/java/com/getstream/sdk/chat/rest/core/Client.java
@@ -56,6 +56,7 @@ import com.getstream.sdk.chat.rest.request.AddDeviceRequest;
 import com.getstream.sdk.chat.rest.request.AddMembersRequest;
 import com.getstream.sdk.chat.rest.request.BanUserRequest;
 import com.getstream.sdk.chat.rest.request.ChannelQueryRequest;
+import com.getstream.sdk.chat.rest.request.GuestUserRequest;
 import com.getstream.sdk.chat.rest.request.MarkReadRequest;
 import com.getstream.sdk.chat.rest.request.QueryChannelsRequest;
 import com.getstream.sdk.chat.rest.request.QueryUserRequest;
@@ -80,6 +81,7 @@ import com.getstream.sdk.chat.rest.response.MuteUserResponse;
 import com.getstream.sdk.chat.rest.response.QueryChannelsResponse;
 import com.getstream.sdk.chat.rest.response.QueryUserListResponse;
 import com.getstream.sdk.chat.rest.response.SearchMessagesResponse;
+import com.getstream.sdk.chat.rest.response.TokenResponse;
 import com.getstream.sdk.chat.rest.response.WsErrorMessage;
 import com.getstream.sdk.chat.rest.storage.BaseStorage;
 import com.getstream.sdk.chat.storage.Storage;
@@ -133,12 +135,14 @@ public class Client implements WSResponseHandler {
 
     private ApiClientOptions apiClientOptions;
     private Boolean offlineStorage;
+    private boolean anonymousConnection;
     private ApiServiceProvider apiServiceProvider;
     private WebSocketServiceProvider webSocketServiceProvider;
     private UploadStorageProvider uploadStorageProvider;
     private StorageProvider storageProvider;
     private Context context;
     private Handler delayedDisconnectWebSocketHandler = new Handler();
+
 
     // Client params
     private Map<String, Channel> activeChannelMap = new HashMap<>();
@@ -157,6 +161,7 @@ public class Client implements WSResponseHandler {
     private int webSocketDisconnectDelay;
 
     private Map<String, Config> channelTypeConfigs;
+
     // endregion
     private ChatEventHandler builtinHandler =
 
@@ -165,7 +170,7 @@ public class Client implements WSResponseHandler {
                 public void onAnyEvent(Event event) {
                     // if an event contains the current user update it
                     // this also captures notification.mutes_updated
-                    if (event.getMe() != null) {
+                    if (event.getMe() != null && !event.isAnonymous()) {
                         state.setCurrentUser(event.getMe());
                     }
                     if (event.getType() == EventType.NOTIFICATION_MUTES_UPDATED) {
@@ -389,6 +394,10 @@ public class Client implements WSResponseHandler {
         return apiService;
     }
 
+    public boolean isAnonymousConnection() {
+        return anonymousConnection;
+    }
+
     public boolean isConnected() {
         return connected;
     }
@@ -450,6 +459,8 @@ public class Client implements WSResponseHandler {
      * @param provider the Token Provider used to obtain the auth token for the user
      */
     public synchronized void setUser(@NotNull User user, @NotNull final TokenProvider provider) {
+        anonymousConnection = false;
+
         if (getUser() != null) {
             Log.w(TAG, "setUser was called but a user is already set; this is probably an integration mistake");
             return;
@@ -497,7 +508,7 @@ public class Client implements WSResponseHandler {
                 cacheUserToken = null;
             }
         };
-        connect();
+        connect(anonymousConnection);
     }
 
     /**
@@ -597,18 +608,28 @@ public class Client implements WSResponseHandler {
         }
     }
 
-    private synchronized void connect() {
+    private synchronized void connect(boolean anonymousConnection) {
         Log.i(TAG, "client.connect was called");
-        tokenProvider.getToken(userToken -> {
+
+        if (anonymousConnection) {
             try {
-                webSocketService = webSocketServiceProvider.provideWebSocketService(getUser(), userToken, this);
-                apiService = apiServiceProvider.provideApiService(tokenProvider);
-                uploadStorage = uploadStorageProvider.provideUploadStorage(tokenProvider, this);
-                webSocketService.connect();
+                webSocketService = webSocketServiceProvider.provideWebSocketService(getUser(), null, this, anonymousConnection);
             } catch (UnsupportedEncodingException e) {
                 onError(e.getMessage(), ClientErrorCode.JSON_ENCODING);
             }
-        });
+        } else {
+            tokenProvider.getToken(token -> {
+                try {
+                    webSocketService = webSocketServiceProvider.provideWebSocketService(getUser(), token, this, anonymousConnection);
+                } catch (UnsupportedEncodingException e) {
+                    onError(e.getMessage(), ClientErrorCode.JSON_ENCODING);
+                }
+            });
+        }
+
+        apiService = apiServiceProvider.provideApiService(tokenProvider, anonymousConnection);
+        uploadStorage = uploadStorageProvider.provideUploadStorage(tokenProvider, this);
+        webSocketService.connect();
     }
 
     public Channel channel(String cid) {
@@ -635,7 +656,7 @@ public class Client implements WSResponseHandler {
     @Override
     public void connectionResolved(Event event) {
         clientID = event.getConnectionId();
-        if (event.getMe() != null)
+        if (event.getMe() != null && !event.isAnonymous())
             state.setCurrentUser(event.getMe());
 
         // mark as connect, any new callbacks will automatically be executed
@@ -694,7 +715,8 @@ public class Client implements WSResponseHandler {
             return;
         }
         connectionRecovered();
-        connect();
+
+        connect(anonymousConnection);
     }
 
     @Override
@@ -1699,12 +1721,48 @@ public class Client implements WSResponseHandler {
      * Setup an anonymous session
      */
     public void setAnonymousUser() {
+        if (getUser() != null) {
+            Log.w(TAG, "setAnonymousUser was called but a user is already set;");
+            return;
+        }
+
+        anonymousConnection = true;
+        apiService = apiServiceProvider.provideApiService(null, true);
+
+        String uuid = randomUUID().toString();
+
+        state.setCurrentUser(new User(uuid));
+
+        connect(anonymousConnection);
     }
 
     /**
      * Setup a temporary guest user
+     *
+     * @param user Data about this user. IE {name: "john"}
      */
     public void setGuestUser(User user) {
+        if (getUser() != null) {
+            Log.w(TAG, "setGuestUser was called but a user is already set;");
+            return;
+        }
+
+        GuestUserRequest body = new GuestUserRequest(user.getId(), user.getName());
+
+        apiService = apiServiceProvider.provideApiService(null, true);
+        apiService.setGuestUser(apiKey, body).enqueue(new Callback<TokenResponse>() {
+            @Override
+            public void onResponse(Call<TokenResponse> call, Response<TokenResponse> response) {
+                if (response.body() != null) {
+                    setUser(user, response.body().getToken());
+                }
+            }
+
+            @Override
+            public void onFailure(Call<TokenResponse> call, Throwable t) {
+                Log.e(TAG, "Problem with setting guest user: " + t.getMessage());
+            }
+        });
     }
 
     /**